# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.ResponseDataSuccess:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: boolean
    type: object
  response.ResponseErr:
    properties:
      code:
        type: integer
      error: {}
      message:
        type: boolean
    type: object
...
  topic_reference_dto.AdminCreateTopicReferenceInput:
    properties:
      name:
        type: string
      path:
        type: string
      teacher_id:
        type: integer
    required:
    - name
    - path
    - teacher_id
    type: object
  topic_reference_dto.TeacherCreateTopicReferenceInput:
    properties:
      name:
        type: string
      path:
        type: string
    required:
    - name
    - path
    type: object
  topic_reference_dto.TeacherUpdateTopicReferenceInput:
    properties:
      id:
        type: integer
      name:
        type: string
      path:
        type: string
    required:
    - id
    - name
    - path
    type: object
...
host: localhost:8080
info:
  contact:
    email: minhbeardev@gmail.com
  description: Capstone Project FPT API
  title: Capstone Project FPT API
  version: "1.0"
paths:
...
  /topic-references/admins:
    post:
      consumes:
      - application/json
      description: Admin create topic reference
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/topic_reference_dto.AdminCreateTopicReferenceInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseDataSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseErr'
      security:
      - ApiKeyAuth: []
      summary: AdminCreateTopicReference
      tags:
      - topic reference
  /topic-references/teachers:
    post:
      consumes:
      - application/json
      description: Teacher create topic reference
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/topic_reference_dto.TeacherCreateTopicReferenceInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseDataSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseErr'
      security:
      - ApiKeyAuth: []
      summary: TeacherCreateTopicReference
      tags:
      - topic reference
    put:
      consumes:
      - application/json
      description: Teacher update topic reference
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/topic_reference_dto.TeacherUpdateTopicReferenceInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseDataSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseErr'
      security:
      - ApiKeyAuth: []
      summary: TeacherUpdateTopicReference
      tags:
      - topic reference
  /topic-references/teachers/{id}:
    delete:
      consumes:
      - application/json
      description: Teacher delete topic reference
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseDataSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseErr'
      security:
      - ApiKeyAuth: []
      summary: TeacherDeleteTopicReference
      tags:
      - topic reference
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"